pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "windvox"
version = "0.1.0"
description = "Linux voice input service using Volcengine Doubao ASR"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "WindVox" }
]

dependencies = [
    "sounddevice>=0.4.6",
    "numpy>=1.24",
    "websockets>=12.0",
    "pynput>=1.7.6",
    "pystray>=0.19",
    "pillow>=10.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
windvox = "windvox.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/windvox"]
